fix: add custom hooks
